hotfix: 채팅 관련 핫픽스 + 채팅 소켓 연결 끊김 방지

See merge request s12-fintech-finance-sub1/S12P21A601!102

+++ back-end/moba/src/main/java/com/a601/moba/chat/Entity/ChatMessage.java
@@ -9,13 +9,13 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public record ChatMessage(
         @Id String id,
         Integer appointmentId,
-        Long senderId,
+        Integer senderId,
         String senderName,
         String message,
         LocalDateTime sentAt
 ) {
     @Builder
-    public ChatMessage(String id, Integer appointmentId, Long senderId, String senderName, String message,
+    public ChatMessage(String id, Integer appointmentId, Integer senderId, String senderName, String message,
                        LocalDateTime sentAt) {
         this.id = id;
         this.appointmentId = appointmentId;
